Move most rotting code to shared
